refactor: split tests to separate file, add missing tests, use f64 accumulation - Extract test module to ngram_tests.rs (ngram.rs: 304 lines, tests: 382 lines) - Add 5 missing tests: whitespace-only, multiple tokens, tabs, CJK query - Use f64 intermediate accumulation for border weight multiplication
